Continuing to add on_error and compile_opt
Continuing to add on_error and compile_opt commands, and have included a
check for a system variable. If system variable !debug exists then the
on_error is skipped, so its easier to debug. Adding the compile_opt,
idl2 caused an error where two structs were compared one used an integer
and the other a long, so comparison failed.